+++ test/audits/conversion.test.js
@@ -11,15 +11,14 @@
  */
 
 /* eslint-env mocha */
-import chai from 'chai';
+import { expect, use } from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 import nock from 'nock';
 import { conversionAuditRunner } from '../../src/conversion/handler.js';
 import { rumDashboardData, rumSourcesData, expectedConversionData } from '../fixtures/conversion-data.js';
 
-chai.use(sinonChai);
-const { expect } = chai;
+use(sinonChai);
 
 const DOMAIN_REQUEST_DEFAULT_PARAMS = {
   interval: 7,